feat: support toggling balance, avatar and chain visibility
